refactor: made code more organized
